more cleanups

+++ modules/base/keyboard.c
@@ -104,7 +104,7 @@ static unsigned int clone_open(struct system_node *self)
     system_group_add(&snode[index].base, &snode[index].device.node);
     system_group_add(&snode[index].base, &snode[index].control.node);
 
-    return (unsigned long)&snode[index].control;
+    return (unsigned int)&snode[index].control;
 
 }
